% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofIntegrations.R
\name{add.index.range}
\alias{add.index.range}
\title{adds mass indices to mass list}
\usage{
add.index.range(mass.list, curr.mass.cal)
}
\description{
adds mass indices to mass list
}
\examples{
mass.list <- read.table(text='
 compound  center
 H3O+  21.022
 acetone   59.049
 isoprene  69.0699',
stringsAsFactor=FALSE, header=TRUE)
mass.list <- create.mass.bounds(mass.list)
add.index.range(mass.list, curr.mass.cal)
}

