% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse}
\alias{analyse}
\title{Produce an html analysis of a ToF file}
\usage{
analyse(tof_file, tune = FALSE)
}
\arguments{
\item{tof_file}{location of the ToF file}

\item{tune}{(optional logical, default FALSE) whether to tune hyper-parameters.
If TRUE, will take a LOT more compute time.}
}
\description{
Given a ToF file, train a variety of classifiers & look at cross-validation
accuracy and feature importance. Classes are determined by the first letter
of each sample name.
}
