% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofh5-file.R
\docType{methods}
\name{close}
\alias{close}
\alias{close,TofH5-method}
\title{close TOF file}
\usage{
close(object)

\S4method{close}{TofH5}(object)
}
\description{
close usually not required
}

