% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofIntegrations.R
\name{create.mass.bounds}
\alias{create.mass.bounds}
\title{adds mass bounds to mass list}
\usage{
create.mass.bounds(mass.list, res = 1500)
}
\description{
adds mass bounds to mass list
}
\examples{
mass.list <- read.table(text='
 compound  center
 H3O+  21.022
 acetone   59.049
 isoprene  69.0699',
stringsAsFactor=FALSE, header=TRUE)
create.mass.bounds(mass.list)
}

