% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TofMeasurementClass.R
\docType{methods}
\name{createTofMeasurement}
\alias{createTofMeasurement}
\title{constructor for TofMeasurement objects}
\usage{
createTofMeasurement(file.path('testdata',"Ac just exhale (2014-10-23T10h43m42_#).h5"))
}
\arguments{
\item{file}{the name of the hdf5 file containing the scan}
}
\value{
a TofMeasurement object representing the file
}
\description{
constructor for TofMeasurement objects
}

