% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidate.R
\name{crossvalidate}
\alias{crossvalidate}
\title{Cross-validate on a matrix of ToF data with labels}
\usage{
crossvalidate(data_matrix, labels, model = "xgbTree", n_folds = 10,
  tune = FALSE, hyperparams = "tof")
}
\arguments{
\item{labels}{A character vector of class labels (shoulkd have 2 unique values)}

\item{model}{The model to use}

\item{n_folds}{number of folds in crossvalidation}

\item{tune}{Whether to tune hyper-parameters (TRUE), or use defaults (FALSE)}

\item{hyperparams}{One of 'tof' or 'gcims' whether the default hyper-
parameters have been picked for tof data or gc-ims data. No effect if tune
= TRUE.}

\item{data_marix}{A matrix of ToF data}
}
\value{
A caret::train object
}
\description{
Perform 10-fold cross-validation on a matrix of ToF data. This produces
a 'caret::train' object containing lots of relevant information. One can
extract predictions made and use these to determine accuracy.
}
\details{
The caret::train object also includes a classifier trained on the full
dataset. #' This can be used to make further predictions on new data, such as
a validation set.
}
