% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidate.R
\name{crossvalidation_feature_importance}
\alias{crossvalidation_feature_importance}
\title{Crossvalidation feature importance}
\usage{
crossvalidation_feature_importance(classifier, n_features = 0,
  override_names = character(0), output_dataframe = FALSE)
}
\arguments{
\item{classifier}{Output from 'crossvalidate'}

\item{n_features}{The number of features to show}

\item{override_names}{A character vector to override the names of the
features. This may be useful if the variable names are too long.}

\item{output_dataframe}{Default FALSE. If set to TRUE, this function returns
a data frame of feature iportances instead of a plot.}
}
\value{
Either a plot or a data frame of feature importance.
}
\description{
Given a classifier output from 'crossvalidate', return feature importances
as either a plot or a data frame. Feature importances are calculated using
the caret::varImp function, which computes importances in a different way
depending on the classifier. I highly recommend reading the documentation
here: https://topepo.github.io/caret/variable-importance.html
}
