% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidate.R
\name{crossvalidation_metrics}
\alias{crossvalidation_metrics}
\title{Crossvalidation metric}
\usage{
crossvalidation_metrics(classifier)
}
\arguments{
\item{classifier}{Output of 'crossvalidate'}
}
\value{
data frame of metrics
}
\description{
Accuracy metrics for the cross-validated predictions made in 'crossvalidate'
}
