% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidate.R
\name{crossvalidation_roc}
\alias{crossvalidation_roc}
\title{Crossvalidation ROC curve}
\usage{
crossvalidation_roc(classifier, title = "")
}
\arguments{
\item{classifier}{Output from 'crossvalidate' function}

\item{title}{Title of the output plot}
}
\value{
a ROC curve
}
\description{
Produce a ROC curve based on the output from the 'crossvalidate' function.
}
