% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop-nearly-all-zero.R
\name{drop_nearly_all_zero}
\alias{drop_nearly_all_zero}
\title{Drop columns in a matrix which are nearly all zeroes}
\usage{
drop_nearly_all_zero(mat, max_zero_proportion = 0.5)
}
\arguments{
\item{mat}{matrix of data}

\item{max_zero_proportion}{value in [0,1]. Colums with more than this
proportion of zeroes are dropped}
}
\value{
the original matrix with nearly-zero columns dropped
}
\description{
Given a matrix, drops all columns which contain more than 'threshold'
proportion of zeroes.
}
