% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/massCalFunctions.R
\name{enrich.ions.block}
\alias{enrich.ions.block}
\title{add variables to ions.block}
\usage{
enrich.ions.block(pos.block)
}
\description{
adds power 1/2, 3/2 and 5/2 of the masses to the position block in order for fitting the more complex mass calibration function of Muller 2013 [1]
}
\references{
[1] M. Müller, T. Mikoviny, W. Jud, B. D’Anna, A. Wisthaler, and B. D’Anna, “A New Software Tool for the Analysis of High Resolution PTR-TOF Mass Spectra,” Chemom. Intell. Lab. Syst., vol. 127, pp. 158–165, Jun. 2013.
}

