% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{extract_labels}
\alias{extract_labels}
\title{Extract two-class labels from a set of sample names}
\usage{
extract_labels(names)
}
\arguments{
\item{names}{A set of sample names}
}
\value{
Character vector of labels
}
\description{
Given a set of names, find the shortest prefix giving two unique classes (or
fail if no such prefix exists). Return a lenth n character vector containing
the classes.
}
\details{
The output is run through make.names() which may add/remove characters, such
that the resultant labels are valid variables names (to play nice with
caret)
}
