% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofh5-calib-ionblock.R
\docType{methods}
\name{find.ion.block}
\alias{find.ion.block}
\alias{find.ion.block,TofH5,data.frame-method}
\title{Calibrate TOF file mass axis}
\usage{
find.ion.block(object, ions, ...)

\S4method{find.ion.block}{TofH5,data.frame}(object, ions,
  preliminary.coeff = init.mass.calib(object), n.samp = 100,
  num.knots = 3)
}
\description{
performs an elaborate mass calibration
}

