% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofReader.R
\name{find_ion_tof}
\alias{find_ion_tof}
\title{find location of ion maximum in TOF spectrum}
\usage{
find_ion_tof(ion.mass = 21.0220875, preliminary.coeff, curr.spec.line,
  wid = 0.3)
}
\description{
find location of ion maximum in TOF spectrum
}

