% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/massCalFunctions.R
\name{fit.mass.calib}
\alias{fit.mass.calib}
\title{fit mass calibration function}
\usage{
fit.mass.calib(curr.h5, pos.block, num.knots = 3)
}
\arguments{
\item{curr.h5}{current measurement}

\item{pos.block}{enriched position data of ions}

\item{num.knots}{number of knots for spline}
}
\value{
linear model fit
}
\description{
uses the more elaborate function described in Muller 2013 [1]
}
\references{
[1] M. Müller, T. Mikoviny, W. Jud, B. D’Anna, A. Wisthaler, and B. D’Anna, “A New Software Tool for the Analysis of High Resolution PTR-TOF Mass Spectra,” Chemom. Intell. Lab. Syst., vol. 127, pp. 158–165, Jun. 2013.
}

