% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofReader.R
\name{get.raw.tofblock}
\alias{get.raw.tofblock}
\title{first step of reading - get the tof block
not export}
\usage{
get.raw.tofblock(fid)
}
\description{
first step of reading - get the tof block
not export
}
\examples{
fid <-H5Fopen(tof.h5)
tofblock <- get.raw.tofblock(fid)
H5close(fid)
}

