% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofHolder.R
\name{h5.holder}
\alias{h5.holder}
\title{list to hold functions for a single H5 measurement}
\usage{
h5.holder(tof.h5)
}
\arguments{
\item{tof.h5}{file path to relevant H5 file}
}
\value{
list of functions
}
\description{
list to hold functions for a single H5 measurement
}
\examples{
\dontrun{
tof.h5 <- file.path('data/',"2015.07.17-10h40m34 Ethanol deurated Karl .h5")
curr.h5 <- h5.holder(tof.h5)
}
}

