% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofh5-mass-calib.R
\docType{methods}
\name{init.mass.calib}
\alias{init.mass.calib}
\alias{init.mass.calib,TofH5-method}
\title{Read inital guesses for mass calibration}
\usage{
init.mass.calib(object, ...)

\S4method{init.mass.calib}{TofH5}(object)
}
\description{
read from the H5 file the calibrated value from Ionicon
}

