% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofIntegrations.R
\name{integrate.full.spec}
\alias{integrate.full.spec}
\title{integrates the full spec}
\usage{
integrate.full.spec(curr.reader, indexhelp, mass.cal, mass.list)
}
\description{
integrates the full spec
}
\examples{
fid <-H5Fopen(tof.h5)
tofblock <- get.raw.tofblock(fid)
indexhelp <- tof.indexhelp(tofblock)
curr.reader <- make.curr.tofreader(tofblock, indexhelp)
mass.list <- read.table(text='
 compound  center
 H3O+  21.022
 acetone   59.049
 isoprene  69.0699',
stringsAsFactor=FALSE, header=TRUE)
mass.list <- create.mass.bounds(mass.list)
int.res <- integrate.full.spec(curr.reader, indexhelp, smooth.mc, mass.list)
library(ggplot2)
ggplot(int.res \%>\% filter(index >1000 & index < 1300)) + geom_line(mapping=aes(x=index, y=area, color=compound)) + scale_y_log10()
}

