% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofIntegrations.R
\name{integrate.one.spec}
\alias{integrate.one.spec}
\title{integrate single spec line}
\usage{
integrate.one.spec(mass.list, spec)
}
\description{
integrate single spec line
}
\examples{
mass.list <- read.table(text='
 compound  center
 H3O+  21.022
 acetone   59.049
 isoprene  69.0699',
stringsAsFactor=FALSE, header=TRUE)
mass.list <- create.mass.bounds(mass.list)
mass.list <- add.index.range(mass.list, curr.mass.cal)
integrate.one.spec(mass.list2, spec)
}

