% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofReader.R
\docType{data}
\name{ions}
\alias{ions}
\title{sample ions for TOF calibration}
\format{\preformatted{ Named num [1:8] 21 30 32 59 39 ...
 - attr(*, "names")= chr [1:8] "h3o" "no" "o2" "aceton" ...
}}
\usage{
ions
}
\description{
sample ions for TOF calibration
}
\keyword{datasets}

