% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-tof.R
\name{load_tof}
\alias{load_tof}
\title{Load ToF data}
\usage{
load_tof(tof_file)
}
\arguments{
\item{tof_file}{location of the ToF file}
}
\value{
A data frame with samples in rows, molecules in columns.
}
\description{
This function takes the location of a ToF file (assumed to be stored as
molecules in columns and samples in rows), and returns a data frame.
}
