% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/massCalFunctions.R
\name{locate.ions.block}
\alias{locate.ions.block}
\title{locate ions for mass calibration}
\usage{
locate.ions.block(df.ion, curr.h5, preliminary.coeff = list(intercept = 900,
  square_mass = 17650), n.samp = 100)
}
\arguments{
\item{df.ion}{named list of ion masses}

\item{curr.h5}{current tof measurement holder}

\item{n.samp}{number of points to take along the measurement}

\item{perliminary.coef}{starting values for coefficients}
}
\value{
data.frame having columns scan, ion, and located position
}
\description{
searches for dominant ions in the relevant places and return a structure for mass calibration.
This should be improved soon to incorporate data alreade precalculated
}

