% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/massCalFunctions.R
\name{locate.ions.block_}
\alias{locate.ions.block_}
\title{locate ions for mass calibration}
\usage{
locate.ions.block_(df.ion, scl, read.scan, preliminary.coeff)
}
\arguments{
\item{df.ion}{named list of ion masses}

\item{scl}{scans to mass calibrate}

\item{perliminary.coef}{starting values for coefficients}

\item{n.samp}{number of points to take along the measurement}
}
\value{
data.frame having columns scan, ion, and located position
}
\description{
searches for dominant ions in the relevant places and return a structure for mass calibration.
This should be improved soon to incorporate data alreade precalculated
}

