% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofReader.R
\name{make.curr.tofreader}
\alias{make.curr.tofreader}
\title{helper to wrap reader call
not export}
\usage{
make.curr.tofreader(tofblock, indexhelp)
}
\description{
helper to wrap reader call
not export
}
\examples{
fid <-H5Fopen(tof.h5)
tofblock <- get.raw.tofblock(fid)
indexhelp <- tof.indexhelp(tofblock)
cr <- make.curr.tofreader(tofblock, indexhelp)
spec5 <- cr(5)
}

