% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/massCalFunctions.R
\name{make.read.block}
\alias{make.read.block}
\title{prepare for index calculations of block}
\usage{
make.read.block(curr.h5, masslist, scans = seq(curr.h5$num.scan()))
}
\arguments{
\item{curr.h5}{the h5 for which the calculation is to be performed}

\item{masslist}{masses to be considered}
}
\value{
data.frame of prepared variables
}
\description{
given a required list of ion masses the relevant indices at given scan is calculated
}
\examples{
\dontrun{
masslist <- seq(46.9,47.1, by=0.001)
make.read.block(curr.h5,masslist)
}
}

