% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofh5-mass-calib.R
\docType{methods}
\name{mass.calib}
\alias{mass.calib}
\alias{mass.calib,TofH5,data.frame-method}
\title{Calibrate TOF file mass axis}
\usage{
mass.calib(object, ions, ...)

\S4method{mass.calib}{TofH5,data.frame}(object, ions,
  preliminary.coeff = init.mass.calib(object), n.samp = 100,
  num.knots = 3)
}
\description{
performs an elaborate mass calibration
}

