% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofReader.R
\name{mass.calib.coeff.single}
\alias{mass.calib.coeff.single}
\title{#calibrate a single spec line}
\usage{
mass.calib.coeff.single(ions, preliminary.coeff, curr.spec.line)
}
\description{
#calibrate a single spec line
}

