% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mass.calib.tof}
\alias{mass.calib.tof}
\title{calculates mass calibration coefficients for every scan in the flat tof array}
\usage{
mass.calib.tof(flat.tof, preliminary.coeff = list(intercept = 900, square_mass
  = 17650), n.ions = unlist(ions))
}
\arguments{
\item{flat.tof}{array as obtained from to.flat.tof}

\item{preliminary.coeff}{some start values for coefficients}

\item{n.ions}{ion list to be used for calibration}
}
\value{
data frame of calibration coefficients
}
\description{
calculates mass calibration coefficients for every scan in the flat tof array
}
\examples{
mc.table <- mass.calib.tof(flat.tof)
}

