% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofh5-file.R
\docType{methods}
\name{open}
\alias{open}
\alias{open,TofH5-method}
\title{open TOF file}
\usage{
open(object)

\S4method{open}{TofH5}(object)
}
\description{
open required for reading tof data
}

