% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{pars.approx}
\alias{pars.approx}
\title{approximate mass calibration}
\format{\preformatted{List of 2
 $ intercept: num 900
 $ slope    : num 17600
}}
\usage{
pars.approx
}
\description{
approximate mass calibration
}
\keyword{datasets}

