% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gc-ims.R
\name{plot_gcims}
\alias{plot_gcims}
\title{Plot one or more GC-IMS matrices}
\usage{
plot_gcims(gcims, names = character(0), subsample = 1,
  max_signal = Inf, nrow = NULL, ncol = NULL,
  ims_transformation = "identity")
}
\arguments{
\item{gcims}{A 'gcims' matrix, or a list of such matrices}

\item{names}{optional character vector to give titles when plotting multiple matrices}

\item{subsample}{(integer) Sub-sample 1-in-`subsample' pixels}

\item{max_signal}{maximum signal plotted (higher vals are capped)}

\item{nrow}{number of rows when plotting multiple matrices}

\item{ncol}{number of cols when plotting multiple matrices}

\item{ims_transformation}{optional transformation to apply (passed to scale fill 'trans' argument). "log1p" is useful.}
}
\value{
ggplot object plotting the matrices
}
\description{
Plot one or more GC-IMS matrices
}
