% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/massCalFunctions.R
\name{predic.positions}
\alias{predic.positions}
\title{predict position index of masses}
\usage{
predic.positions(model, read.block)
}
\arguments{
\item{fitted}{model}

\item{required}{masses}
}
\value{
data.frame of index position
}
\description{
uses the parameters of readblock to predict index of specified masses
}
\examples{
\dontrun{
masslist <- seq(46.9,47.1, by=0.001)
rmod <- fit.mass.calib(curr.h5, pos.block)
rb <- make.read.block(curr.h5,masslist)
pred.pos <- predic.positions(rmod, rb)
}
}

