% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofIntegrations.R
\name{process.tof.h5}
\alias{process.tof.h5}
\title{Processes full tofwerks H5 file}
\usage{
process.tof.h5(tof.h5, mass.list, n.ions = unlist(ions))
}
\description{
Processes full tofwerks H5 file
}
\examples{
tof.h5 <- file.path('data',"Ac just breath after C (2014-10-23T11h34m53_#).h5")
tof.h5 <- file.path('data',"uncal.h5")
mass.list <- read.table(text='
                        compound  center
                        H3O+  21.022
                        acetone   59.049
                        isoprene  69.0699',
                        stringsAsFactor=FALSE, header=TRUE)
mass.list <- create.mass.bounds(mass.list)
res <- process.tof.h5(tof.h5, mass.list)
ggplot(res \%>\% filter(index >1000 & index < 1300)) + geom_line(mapping=aes(x=index, y=area, color=compound)) + scale_y_log10()
}

