% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofh5-scan.R
\docType{methods}
\name{read.bins}
\alias{read.bins}
\alias{read.bins,TofH5-method}
\title{Read TOF scan binned at provided masses}
\usage{
read.bins(object, ...)

\S4method{read.bins}{TofH5}(object, mass.calibration, index, masses)
}
\description{
Read TOF scan binned at provided masses
}

