% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.fullspectra.tof.h5}
\alias{read.fullspectra.tof.h5}
\title{reads the full spectrum matrix from Tofwerks hdf5 file}
\usage{
read.fullspectra.tof.h5(tof.h5)
}
\arguments{
\item{tof.h5}{filename of the hdf5 file}
}
\description{
reads the full spectrum matrix from Tofwerks hdf5 file
}
\examples{
raw.tof <- read.fullspectra.tof.h5(file.path('data',"uncal.h5"))
}

