% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofh5-scan.R
\docType{methods}
\name{read.prepared.bins}
\alias{read.prepared.bins}
\alias{read.prepared.bins,TofH5-method}
\title{Read TOF scan at prepared targets}
\usage{
read.prepared.bins(object, ...)

\S4method{read.prepared.bins}{TofH5}(object, targets)
}
\arguments{
\item{targets}{list of targets with column idx specifying the relevant TOF indices and column index the relevant scan number}
}
\description{
Read TOF scan at prepared targets
}
\examples{
\dontrun{
tof.h5 <- file.path('testdata/Ac just exhale (2014-10-23T10h43m42_#).h5')
data(sample_ions)
ions <- sample_ions

object <- new("TofH5", file.name=tof.h5)
object <- open(object)

scl <- sample.scans(object, n.samp=100)

ion.block <- find.ion.block(object,ions)


ion.block <- enrich.ib(ion.block)
head(ion.block)
num.knots <- 2
kn <- floor(seq(1,to=object@.indexhelp$N, length.out=num.knots+2))[2:num.knots+1]
fit.mos <- rlm(pos ~ (isq + ion + sq) * ns(scan,knots=kn), ion.block, method='MM')

extr.data <- expand.grid(scan= 1:10, ion=seq(47,47.1,length.out=100))
extr.data <- enrich.ib(extr.data)
extr.data$idx <- predict(fit.mos, newdata = extr.data)

read.prepared.bins(object,targets=extr.data)
}
}

