% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofh5-scan.R
\docType{methods}
\name{read.raw.scan}
\alias{read.raw.scan}
\alias{read.raw.scan,TofH5,numeric-method}
\title{Read TOF scan at specified index}
\usage{
read.raw.scan(object, index)

\S4method{read.raw.scan}{TofH5,numeric}(object, index)
}
\description{
Read TOF scan at specified index
}

