% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofReader.R
\name{read.spec.ind}
\alias{read.spec.ind}
\title{reads the spec of a single tof scan}
\usage{
read.spec.ind(tofblock, indexhelp, i)
}
\description{
reads the spec of a single tof scan
}
\examples{
fid <-H5Fopen(tof.h5)
tofblock <- get.raw.tofblock(fid)
indexhelp <- tof.indexhelp(tofblock)
spec <- read.spec.ind(tofblock, indexhelp, 40)
plot(spec,type='l')
}

