% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.tof.fullspectra}
\alias{read.tof.fullspectra}
\title{reads the full spectrum matrix from Tofwerks hdf5 file and converts it to flat}
\usage{
read.tof.fullspectra(tof.h5)
}
\arguments{
\item{tof.h5}{filename of the hdf5 file}
}
\description{
reads the full spectrum matrix from Tofwerks hdf5 file and converts it to flat
}
\examples{
flat.tof <- read.tof.fullspectra(file.path('data',"uncal.h5"))
}

