% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofReader.R
\name{read.tof.peaks}
\alias{read.tof.peaks}
\title{reads integrated peaks as integrated by PTR-MS TOD-DAQ-Viewer.}
\usage{
read.tof.peaks(tof.h5)
}
\arguments{
\item{tof.h5}{filename of hdf5 file}
}
\value{
dataframe with structure: rows = scans, cols = counts + information
}
\description{
WARNING: these must be first integrated in the software otherwise and error
will be thrown.
}
\examples{
tof.h5 <- file.path('data',"uncal.h5")
}

