% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_bin_c}
\alias{read_bin_c}
\title{calculates the number of counts between indices}
\usage{
read_bin_c(ys, pos)
}
\arguments{
\item{ys}{spectrum vector}

\item{pos}{position vector}
}
\value{
vector of length pos with the summed counts (first is 0)
}
\description{
calculates the number of counts between indices
}

