% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofh5-mass-calib.R
\docType{methods}
\name{sample.scans}
\alias{sample.scans}
\alias{sample.scans,TofH5-method}
\title{Sample a number of scans equally over the measurments}
\usage{
sample.scans(object, ...)

\S4method{sample.scans}{TofH5}(object, n.samp = 100)
}
\description{
Sample a number of scans equally over the measurments
}

