% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_data.R
\docType{data}
\name{sample_ions}
\alias{sample_ions}
\title{Example data set of ions used for mass calibration}
\format{\preformatted{'data.frame':	13 obs. of  1 variable:
 $ ions: num  21 39 55 30 32 ...
}}
\usage{
sample_ions
}
\description{
Example data set of ions used for mass calibration
}
\author{
Clemens Ager \email{clemens.ager@uibk.ac.at}
}
\keyword{data}
\keyword{ions}

