% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofReader.R
\name{smooth.mass.cal}
\alias{smooth.mass.cal}
\title{smoothes the mass calibration to avoid local jumps}
\usage{
smooth.mass.cal(mc.table)
}
\arguments{
\item{mc.table}{raw mass calibration}
}
\value{
mass calibration table after smoothing
}
\description{
smoothes the mass calibration to avoid local jumps
}
\examples{
mc.table <- mass.calib.tof(tofblock, indexhelp)
smooth.mc <- smooth.mass.cal(mc.table)
}

