% Generated by roxygen2 (4.0.2): do not edit by hand
\name{to.flat.tof}
\alias{to.flat.tof}
\title{converts the raw multidimensional TOF array into flat structure}
\usage{
to.flat.tof(raw.tof.array)
}
\arguments{
\item{raw.tof.array}{multidimensional array as present in hdf5 file, as obtained by read.fullspectra.tof.h5}
}
\description{
converts the raw multidimensional TOF array into flat structure
}
\examples{
flat.tof <- to.flat.tof(raw.tof)
}

