% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofReader.R
\name{tof.indexhelp}
\alias{tof.indexhelp}
\title{pre calculates some values
not export}
\usage{
tof.indexhelp(tofblock)
}
\description{
pre calculates some values
not export
}
\examples{
fid <-H5Fopen(tof.h5)
tofblock <- get.raw.tofblock(fid)
indexhelp <- tof.indexhelp(tofblock)
}

