% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofh5-class.R
\docType{class}
\name{tofh5-class}
\alias{tofh5-class}
\title{TOF H5 measurement class}
\description{
designed to hold and interact with the H5 measurment
data generated by Ionicon Measurement device. Hopefully this will by usefull for other devices as well.
}
\section{Slots}{

\describe{
\item{\code{file.name}}{holding the HDF5 data}

\item{\code{n.scans}}{number of scans in file for easy access}
}}

