% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TofGetJustScan.R, R/TofGetScan.R
\docType{methods}
\name{getScan}
\alias{getJustScan,TofMeasurement,numeric-method}
\alias{getScan}
\alias{getScan,TofMeasurement,numeric-method}
\title{get just the individual scan (no mass axis included)}
\usage{
\S4method{getJustScan}{TofMeasurement,numeric}(object, index = 1)

\S4method{getScan}{TofMeasurement,numeric}(object, index = 1)
}
\arguments{
\item{object}{the TofMeasurement object}

\item{index}{1 based index of scan}

\item{object}{the TofMeasurement object}

\item{index}{1 based index of scan}
}
\value{
MassSpectrum object from MALDIquant

MassSpectrum object from MALDIquant
}
\description{
this uses a continous index instead of the buf-write notation

this uses a continous index instead of the buf-write notation
}

