% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TofMeasurementClass.R
\docType{class}
\name{tofmeasurement}
\alias{TofMeasurement}
\alias{tofmeasurement}
\title{Class TofMeasurement}
\description{
Class \code{TofMeasurement} handles access to a Tof hdf5 file
}
\section{Slots}{

\describe{
\item{\code{file}}{Name of the hdf5 file}

\item{\code{metaDataScan}}{some extra informations}
}}

