% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofTools.R
\docType{package}
\name{tofTools-package}
\alias{tofTools}
\alias{tofTools-package}
\title{Tools to work with PTR-TOF data}
\description{
Tools to work with PTR-TOF data
}
\details{
The data produced by Ionicon PTR-TOF-MS is
in a hdf5 file.  This package aims to provide tools
to directly use this data in R
...
}
\author{
\email{clemens.ager@uibk.ac.at}
}
\seealso{
\code{\link{...}}
}
\keyword{package}

