% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc-ims.R
\name{vec_to_mat}
\alias{vec_to_mat}
\title{GCIMS vector to matrix}
\usage{
vec_to_mat(gcims_vec)
}
\arguments{
\item{gcims_vec}{a names vector cotnaining a gcims measurement}
}
\description{
Convert a GCIMS vector with informative column names into a GCIMS matrix
}
