% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_dir}
\alias{analyse_dir}
\title{Run 'analyse' on every '.txt' file in a directory}
\usage{
analyse_dir(tof_dir, pattern = "*.txt")
}
\arguments{
\item{tof_dir}{directory containing ToF files (names ending in .txt)}

\item{pattern}{The filename patterns to match (default '*.txt')}
}
\value{
a list of error messages; one per file.
}
\description{
Run 'analyse' on every '.txt' file in a directory
}
