% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_GCIMS}
\alias{analyse_GCIMS}
\title{Produce a report on the predictive accuracy of a directory of pre-processed
GC-IMS data.}
\usage{
analyse_GCIMS(dir)
}
\arguments{
\item{dir}{directory containing pre-processed GC-IMS files}
}
\description{
Produce a report on the predictive accuracy of a directory of pre-processed
GC-IMS data.
}
