% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ions}
\alias{ions}
\title{sample ions for TOF calibration}
\format{\preformatted{List of 6
 $ h3o    : num 21
 $ no     : num 30
 $ aceton : num 59
 $ isopren: num 69.1
 $ h5o2   : num 39
 $ h7o3   : num 55
}}
\usage{
ions
}
\description{
sample ions for TOF calibration
}
\keyword{datasets}

