% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofReader.R
\name{mass.calib.tof}
\alias{mass.calib.tof}
\title{calculates mass calibration coefficients for every scan in the flat tof array}
\usage{
mass.calib.tof(tofblock, indexhelp, preliminary.coeff = list(intercept = 900,
  square_mass = 17650), n.ions = unlist(ions))
}
\arguments{
\item{tofblock}{from get.raw.tofblock}

\item{indexhelp}{from tof.indexhelp}

\item{preliminary.coeff}{some start values for coefficients}

\item{n.ions}{ion list to be used for calibration}
}
\value{
data frame of calibration coefficients
}
\description{
calculates mass calibration coefficients for every scan in the flat tof array
}
\note{
this thkes about 30s for a measurement containing 3000 scans
}
\examples{
tof.h5 <- file.path('data',"Ac just breath after C (2014-10-23T11h34m53_#).h5")
fid <-H5Fopen(tof.h5)
tofblock <- get.raw.tofblock(fid)
indexhelp <- tof.indexhelp(tofblock)
cr <- make.curr.tofreader(tofblock, indexhelp)
mc.table <- mass.calib.tof(tofblock, indexhelp)
}

