% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tofReader.R
\docType{data}
\name{pars.approx}
\alias{pars.approx}
\title{approximate mass calibration}
\format{\preformatted{List of 2
 $ intercept  : num 900
 $ square_mass: num 17600
}}
\usage{
pars.approx
}
\description{
approximate mass calibration
}
\keyword{datasets}

