% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gc-ims.R
\name{plot_features}
\alias{plot_features}
\title{Plot features over a reference matrix}
\usage{
plot_features(importance, reference_gcims, n_features = 0,
  reverse_values = FALSE, discretise_features = FALSE, ...)
}
\arguments{
\item{importance}{vector of feature values (importance)}

\item{reference_gcims}{gcims matrix over which to plot the features}

\item{n_features}{(optional) plot only the top n_features}

\item{reverse_values}{(optional, logical) Consider small feature values as
'big' for the purpose of the colour scale and picking the top n_features}

\item{discretise_features}{show a ret point over each feature, instead of a
colour scale}

\item{...}{passed to plot_gcims}
}
\value{
gcims ggplot overlayed with feature values at the correct locations
}
\description{
Takes a vector of features (with the standard <GC>:<IMS> feature names) and
plots the values over a reference gcims matrix. Can be used to show feature
importance/p-values.
}
