% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gc-ims.R
\name{plot_gcims}
\alias{plot_gcims}
\title{Plot one or more GC-IMS matrices}
\usage{
plot_gcims(gcims, names = character(0), subsample = 1,
  max_signal = Inf, nrow = NULL, ncol = NULL,
  ims_transformation = "identity")
}
\arguments{
\item{gcims}{A 'gcims' matrix, or a list of such matrices}

\item{subsample}{(integer) Sub-sample 1-in-`subsample' pixels}
}
\value{
ggplot object plotting the matrices
}
\description{
Plot one or more GC-IMS matrices
}
