% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.tof.peaks}
\alias{read.tof.peaks}
\title{reads integrated peaks as integrated by PTR-MS TOD-DAQ-Viewer.}
\usage{
read.tof.peaks(tof.h5)
}
\arguments{
\item{tof.h5}{filename of hdf5 file}
}
\value{
dataframe with structure: rows = scans, cols = counts + information
}
\description{
WARNING: these must be first integrated in the software otherwise and error
will be thrown.
}

